refactor : [step4] domain 패키지 세분화
